+++ src/resolvers/FacadeResolver.ts
@@ -7,9 +7,9 @@ export class FacadeResolver implements IFacadeResolver {
         this.resolvers = resolvers;
     }
 
-    public async resolve(accessor: string): Promise<FacadeResolution | null> {
+    public async resolve(accessor: string, method?: string): Promise<FacadeResolution | null> {
         for (const resolver of this.resolvers) {
-            const result = await resolver.resolve(accessor);
+            const result = await resolver.resolve(accessor, method);
             if (result) {
                 return result;
             }